WLA-22: add new Language obj for Legal Decs
