+++ src/main/java/uk/gov/hmcts/reform/probate/model/documents/LegalDeclaration.java
@@ -1,64 +1,28 @@
 package uk.gov.hmcts.reform.probate.model.documents;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonRootName;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
 
 @JsonRootName(value = "legalDeclaration")
 public class LegalDeclaration implements BusinessDocument {
 
-    @NotEmpty
-    @JsonProperty("headers")
-    private List<String> headers;
+    Language en;
+    Language cy;
 
-    @NotEmpty
-    @Valid
-    @JsonProperty("sections")
-    private List<DeclarationSection> sections = new ArrayList<>();
-
-    @NotBlank
-    @JsonProperty("dateCreated")
-    private String dateCreated;
-
-    @NotBlank
-    @JsonProperty("deceased")
-    private String deceased;
-
-    public String getDeceased() {
-        return deceased;
+    public Language getEn() {
+        return en;
     }
 
-    public void setDeceased(String deceased) {
-        this.deceased = deceased;
+    public void setEn(Language en) {
+        this.en = en;
     }
 
-    public String getDateCreated() {
-        return dateCreated;
+    public Language getCy() {
+        return cy;
     }
 
-    public void setDateCreated(String dateCreated) {
-        this.dateCreated = dateCreated;
+    public void setCy(Language cy) {
+        this.cy = cy;
     }
 
-    public List<String> getHeaders() {
-        return headers;
-    }
-
-    public void setHeaders(List<String> headers) {
-        this.headers = headers;
-    }
-
-    public List<DeclarationSection> getSections() {
-        return sections;
-    }
-
-    public void setSections(List<DeclarationSection> sections) {
-        this.sections = sections;
-    }
 
 }
